refactor:change role enum
